feat(admin-api): Added a new `revoke_zome_call_capability` endpoint to the Admin API (#5136)
* feat(admin-api): Added a new `revoke_zome_call_capability` endpoint to the Admin API

The new endpoint allows you to revoke a previously granted Zome Capability

closes #4596

* fix(admin-api)!: Changed the type of `AdminResponse::CapabilityGrantsInfo` to a `Vec<(CellId, Vec<CapGrantInfo>)>`

Changed the type of `AdminResponse::CapabilityGrantsInfo` to a `Vec<(CellId, Vec<CapGrantInfo>)>` instead of a `HashMap<CellId, Vec<CapGrantInfo>>`; this is to make it work with msgpack encoding, which does not support maps with non-string tuple keys.

* feat(admin_api)!: the return type of `grant_zome_call_capability` has been changed from `()` to `ActionHash`.

This has been done to make it easier to know the `ActionHash` of the grant in case you want to revoke it later.